requirement.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
mysqlclient==2.0.3
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
